# Conflicts:
#	Tarea4/Client/client.c
#	Tarea4/Client/client.h
#	Tarea4/cmake-build-debug/CMakeFiles/Tarea4.dir/Client/client.c.o
#	Tarea4/cmake-build-debug/CMakeFiles/Tarea4.dir/Config/config.c.o
#	Tarea4/cmake-build-debug/CMakeFiles/Tarea4.dir/INI/ini.c.o
#	Tarea4/cmake-build-debug/CMakeFiles/Tarea4.dir/compiler_depend.make
#	Tarea4/cmake-build-debug/CMakeFiles/Tarea4.dir/main.c.o
#	Tarea4/cmake-build-debug/Tarea4
#	Tarea4/cmake-build-debug/Testing/Temporary/LastTest.log

+++ Tarea4/Client/balls_data.h
@@ -10,8 +10,8 @@ typedef struct {
     char type_message[20]; // Tipo de mensaje (ej. "ball_data")
     bool active; //True o False
     int id;
-    int posx;
-    int posy;
+    float posx;
+    float posy;
 
 } DataBall;
